+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ### 🆕 Features
 
 - Add JSON-RPC client to connect to any JSON-RPC server over HTTP
+- Add Eth1 client to connect to any Ethereum 1.0 node over HTTP
 
 ### 🛠 Bug fixes